Abstract methods to separate timetable class
